+++ mb-markdown-editor/src/components/App/App.js
@@ -4,7 +4,7 @@ import './App.css';
 function App() {
   return (
     <div className="App">
-
+      
     </div>
   );
 }
